Rushed but well working script
I was scanning a bunch of papers as pdfs and thought it would probably also be great to have them as image files for some people to easier click through it...
I took my cleanbox script and modified it. Added pdf2image. I might rewrite it later as I was in a rush this time and I'm already thinking up some changes to make.